+++ packages/webview_flutter/webview_flutter_platform_interface/test/platform_webview_controller_test.dart
@@ -485,6 +485,21 @@ void main() {
   );
 
   test(
+    'Default implementation of setUserInteractionEnabled should throw unimplemented error',
+    () {
+      final PlatformWebViewController controller =
+          ExtendsPlatformWebViewController(
+            const PlatformWebViewControllerCreationParams(),
+          );
+
+      expect(
+        () => controller.setUserInteractionEnabled(true),
+        throwsUnimplementedError,
+      );
+    },
+  );
+
+  test(
     'Default implementation of setOnJavaScriptAlertDialog should throw unimplemented error',
     () {
       final PlatformWebViewController controller =